+++ src/DTO/IncomeServiceItem.php
@@ -3,6 +3,8 @@ declare(strict_types=1);
 
 namespace Shoman4eg\Nalog\DTO;
 
+use Brick\Math\BigDecimal;
+
 /**
  * @author Artem Dubinin <artem@dubinin.me>
  */
@@ -30,5 +32,33 @@ final class IncomeServiceItem implements \JsonSerializable
             'amount' => $this->amount,
             'quantity' => $this->quantity,
         ];
+    }
+
+    public function getName(): string
+    {
+        return $this->name;
+    }
+
+    /**
+     * @return float|int|string
+     */
+    public function getAmount()
+    {
+        return $this->amount;
+    }
+
+    /**
+     * @return float|int
+     */
+    public function getQuantity()
+    {
+        return $this->quantity;
+    }
+
+    public function getTotalAmount(): BigDecimal
+    {
+        return BigDecimal::of($this->amount)
+            ->multipliedBy($this->quantity)
+        ;
     }
 }